feat: add login Service and jwt test
