--- src/main/java/com/aloc/aloc/domain/Algorithm.java
@@ -1,21 +1,0 @@
-package com.aloc.aloc.domain;
-
-import com.aloc.aloc.common.domain.AuditingTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-
-@Entity
-@Getter
-public class Algorithm extends AuditingTimeEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
-
-	@Column(nullable = false)
-	private String name;
-}
