--- src/main/java/com/aloc/aloc/domain/History.java
@@ -1,31 +1,0 @@
-package com.aloc.aloc.domain;
-
-import com.aloc.aloc.common.domain.AuditingTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.Getter;
-
-@Entity
-@Getter
-public class History extends AuditingTimeEntity {
-
-	@Id
-	private String id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id")
-	private User user;
-
-	@Column(nullable = false)
-	private String username;
-
-	@Column(nullable = false)
-	private String icon;
-
-	private Integer rank;
-}
