--- src/main/java/com/aloc/aloc/domain/Problem.java
@@ -1,32 +1,0 @@
-package com.aloc.aloc.domain;
-
-import com.aloc.aloc.common.domain.AuditingTimeEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Lob;
-import lombok.Getter;
-
-@Entity
-@Getter
-public class Problem extends AuditingTimeEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(length = 36)
-	private Integer id;
-
-	@Column(nullable = false)
-	private String title;
-
-	@Column(nullable = false, columnDefinition = "TEXT")
-	private String tags;
-
-	@Column(nullable = false)
-	private Integer difficulty;
-
-	private Integer algorithmId;
-
-}
