+++ src/main/java/com/aloc/aloc/domain/problem/Problem.java
@@ -1,0 +1,31 @@
+package com.aloc.aloc.domain.problem;
+
+import com.aloc.aloc.global.domain.AuditingTimeEntity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+
+@Entity
+@Getter
+public class Problem extends AuditingTimeEntity {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(length = 36)
+	private Integer id;
+
+	@Column(nullable = false)
+	private String title;
+
+	@Column(nullable = false, columnDefinition = "TEXT")
+	private String tags;
+
+	@Column(nullable = false)
+	private Integer difficulty;
+
+	private Integer algorithmId;
+
+}
