+++ src/main/java/com/aloc/aloc/domain/solvedproblem/SolvedProblem.java
@@ -1,0 +1,37 @@
+package com.aloc.aloc.domain.solvedproblem;
+
+import java.time.LocalDateTime;
+
+import com.aloc.aloc.domain.problem.Problem;
+import com.aloc.aloc.domain.user.User;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.Getter;
+
+@Entity
+@Getter
+public class SolvedProblem {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Integer id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private User user;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "problem_id")
+	private Problem problem;
+
+	private LocalDateTime solvedAt;
+
+
+
+
+}
