--- src/main/java/com/aloc/aloc/enums/Authority.java
@@ -1,5 +1,0 @@
-package com.aloc.aloc.enums;
-
-public enum Authority {
-	ROLE_USER, ROLE_ADMIN, ROLE_GUEST
-}
