--- src/main/java/com/aloc/aloc/security/LoginFailureHandler.java
@@ -1,26 +1,0 @@
-package com.aloc.aloc.security;
-
-import java.io.IOException;
-
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class LoginFailureHandler extends SimpleUrlAuthenticationFailureHandler {
-
-	@Override
-	public void onAuthenticationFailure(
-		HttpServletRequest request,
-		HttpServletResponse response,
-		AuthenticationException exception
-	) throws IOException {
-		response.setStatus(HttpServletResponse.SC_UNAUTHORIZED); //401 인증 실패
-		response.getWriter().write("fail");
-		log.info("로그인 실패");
-	}
-}
